pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "planning-ai"
version = "0.1.0"
description = "AI-powered PI Planning Assistant for Agile Teams"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Piotr Piwowarski", email = "piotr.piwowarski@example.com"}
]
keywords = ["agile", "pi-planning", "ai", "azure-devops", "openai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Project Managers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Project Management",
]

dependencies = [
    "streamlit>=1.32.0",
    "pandas>=2.0.0",
    "plotly>=5.18.0",
    "networkx>=3.1",
    "azure-devops>=7.1.0b1",
    "azure-identity>=1.15.0",
    "pydantic>=2.6.0",
    "openai>=1.12.0",
    "pyyaml>=6.0.1",
    "openpyxl>=3.1.2",  # For Excel export
    "requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/planning-ai"
Repository = "https://github.com/yourusername/planning-ai"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Relaxed for now
ignore_missing_imports = true
